backend/requirements.txt
# ── Core ──────────────────────────────────────────────────
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic[email]==2.*
pydantic-settings==2.*
python-multipart==0.0.*

# ── Auth ──────────────────────────────────────────────────
PyJWT==2.*
bcrypt==4.*

# ── Supabase ──────────────────────────────────────────────
supabase==2.*

# ── Vector DB ─────────────────────────────────────────────
qdrant-client==1.*

# ── Embeddings ────────────────────────────────────────────
cohere==5.*

# ── LLM ───────────────────────────────────────────────────
groq==0.*

# ── HTTP client (Docling calls) ──────────────────────────
httpx==0.*

# ── Local document parsing fallback ───────────────────────
pypdf==5.*
python-docx==1.*
